+++ examples/functions/ipv6/variable.tf
@@ -1,5 +1,5 @@
 variable "ipv6_address" {
-  type = number
+  type = string
   validation {
     condition     = provider::assert::ipv6(var.ipv6_address)
     error_message = "Invalid IPv6 address"